+++ server/api/views/technician_view.py
@@ -9,17 +9,18 @@ from base.models import User
 from django.core.mail import send_mail
 from django.conf import settings
 from rolepermissions.roles import assign_role
+from rolepermissions.checkers import has_permission
 
 @api_view(['GET','POST'])
 @permission_classes([IsAuthenticated])
 def apiTechnicians(request):
     # List technicians
-    if request.method == 'GET':
+    if request.method == 'GET' and has_permission(request.user, 'get_technicians'):
         technicians = Technician.objects.all()
         serializer = TechnicianSerializer(technicians, many=True)
         return Response(serializer.data)
     # Create technician
-    elif request.method == 'POST':
+    elif request.method == 'POST' and has_permission(request.user, 'create_technicians'):
         if 'company' not in request.data.keys():
             return Response(status=status.HTTP_400_BAD_REQUEST)
         company = Company.objects.get(pk=request.data['company'])
@@ -47,6 +48,7 @@ def apiTechnicians(request):
             send_mail(subject, message, from_email, [to_email], fail_silently=False)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    return Response("This user cannot perform this action.", status=status.HTTP_401_UNAUTHORIZED)
 
 @api_view(['GET','PUT','DELETE'])
 @permission_classes([IsAuthenticated])
@@ -56,17 +58,18 @@ def apiTechnician(request,pk):
     except Technician.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
     # Detail of technician
-    if request.method == 'GET':
+    if request.method == 'GET' and has_permission(request.user, 'view_technicians'):
         serializer = TechnicianSerializer(technician, many=False)
         return Response(serializer.data)
     # Update technician
-    elif request.method == 'PUT':
+    elif request.method == 'PUT' and has_permission(request.user, 'update_technicians'):
         serializer = TechnicianSerializer(technician, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
     # Delete technician
-    elif request.method == 'DELETE':
+    elif request.method == 'DELETE' and has_permission(request.user, 'delete_technicians'):
         technician.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
+    return Response("This user cannot perform this action.", status=status.HTTP_401_UNAUTHORIZED)
